Update app name, generall refactoring & reorganization
