+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.embrace",
+    package = "com.revery",
 )
 
 android_resource(
     name = "res",
-    package = "com.embrace",
+    package = "com.revery",
     res = "src/main/res",
 )
