+++ android/app/build.gradle
@@ -102,7 +102,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.embrace"
+        applicationId "com.revery"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
@@ -139,6 +139,7 @@ android {
 
 dependencies {
     implementation project(':@react-native-community_async-storage')
+    implementation project(':react-native-vector-icons')
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "com.android.support:appcompat-v7:${rootProject.ext.supportLibVersion}"
     implementation "com.facebook.react:react-native:+"  // From node_modules
