+++ android/app/src/main/java/com/revery/MainActivity.java
@@ -1,4 +1,4 @@
-package com.embrace;
+package com.revery;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "embrace";
+        return "revery";
     }
 }
