+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "embrace",
-  "displayName": "embrace"
+  "name": "revery",
+  "displayName": "revery"
 }
