+++ index.ios.js
@@ -1,7 +1,7 @@
 import { AppRegistry } from 'react-native';
 import App from './App.js';
 
-AppRegistry.registerComponent('embrace', () => App);
+AppRegistry.registerComponent('revery', () => App);
 
 // The below line is necessary for use with the TestBed App
 AppRegistry.registerComponent('ViroSample', () => App);
