+++ ios/revery-tvOS/Info.plist
