+++ ios/revery-tvOSTests/Info.plist
