+++ ios/revery.xcodeproj/xcshareddata/xcschemes/revery-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "embrace-tvOS.app"
-               BlueprintName = "embrace-tvOS"
-               ReferencedContainer = "container:embrace.xcodeproj">
+               BuildableName = "revery-tvOS.app"
+               BlueprintName = "revery-tvOS"
+               ReferencedContainer = "container:revery.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "embrace-tvOSTests.xctest"
-               BlueprintName = "embrace-tvOSTests"
-               ReferencedContainer = "container:embrace.xcodeproj">
+               BuildableName = "revery-tvOSTests.xctest"
+               BlueprintName = "revery-tvOSTests"
+               ReferencedContainer = "container:revery.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "embrace-tvOSTests.xctest"
-               BlueprintName = "embrace-tvOSTests"
-               ReferencedContainer = "container:embrace.xcodeproj">
+               BuildableName = "revery-tvOSTests.xctest"
+               BlueprintName = "revery-tvOSTests"
+               ReferencedContainer = "container:revery.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "embrace-tvOS.app"
-            BlueprintName = "embrace-tvOS"
-            ReferencedContainer = "container:embrace.xcodeproj">
+            BuildableName = "revery-tvOS.app"
+            BlueprintName = "revery-tvOS"
+            ReferencedContainer = "container:revery.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "embrace-tvOS.app"
-            BlueprintName = "embrace-tvOS"
-            ReferencedContainer = "container:embrace.xcodeproj">
+            BuildableName = "revery-tvOS.app"
+            BlueprintName = "revery-tvOS"
+            ReferencedContainer = "container:revery.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "embrace-tvOS.app"
-            BlueprintName = "embrace-tvOS"
-            ReferencedContainer = "container:embrace.xcodeproj">
+            BuildableName = "revery-tvOS.app"
+            BlueprintName = "revery-tvOS"
+            ReferencedContainer = "container:revery.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
