+++ ios/revery.xcodeproj/xcshareddata/xcschemes/revery.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "embrace.app"
-               BlueprintName = "embrace"
-               ReferencedContainer = "container:embrace.xcodeproj">
+               BuildableName = "revery.app"
+               BlueprintName = "revery"
+               ReferencedContainer = "container:revery.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "embraceTests.xctest"
-               BlueprintName = "embraceTests"
-               ReferencedContainer = "container:embrace.xcodeproj">
+               BuildableName = "reveryTests.xctest"
+               BlueprintName = "reveryTests"
+               ReferencedContainer = "container:revery.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "embraceTests.xctest"
-               BlueprintName = "embraceTests"
-               ReferencedContainer = "container:embrace.xcodeproj">
+               BuildableName = "reveryTests.xctest"
+               BlueprintName = "reveryTests"
+               ReferencedContainer = "container:revery.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "embrace.app"
-            BlueprintName = "embrace"
-            ReferencedContainer = "container:embrace.xcodeproj">
+            BuildableName = "revery.app"
+            BlueprintName = "revery"
+            ReferencedContainer = "container:revery.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "embrace.app"
-            BlueprintName = "embrace"
-            ReferencedContainer = "container:embrace.xcodeproj">
+            BuildableName = "revery.app"
+            BlueprintName = "revery"
+            ReferencedContainer = "container:revery.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "embrace.app"
-            BlueprintName = "embrace"
-            ReferencedContainer = "container:embrace.xcodeproj">
+            BuildableName = "revery.app"
+            BlueprintName = "revery"
+            ReferencedContainer = "container:revery.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
