+++ ios/revery/AppDelegate.h
