+++ ios/revery/Images.xcassets/AppIcon.appiconset/Contents.json
