+++ ios/reveryTests/Info.plist
