+++ ios/reveryTests/reveryTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface embraceTests : XCTestCase
+@interface reveryTests : XCTestCase
 
 @end
 
-@implementation embraceTests
+@implementation reveryTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
