+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "embrace",
+  "name": "revery",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -13,6 +13,7 @@
     "react": "16.8.3",
     "react-native": "^0.59.3",
     "react-native-ui-kitten": "^4.1.1",
+    "react-native-vector-icons": "^6.6.0",
     "react-viro": "2.16.0"
   },
   "devDependencies": {
